feat(dashboard): refactor AI suggestions handling and introduce custom hook
- Refactored AI suggestions logic by creating a new custom hook, `useAiSuggestions`, to encapsulate AI-related functionalities and improve code organization.
- Updated the `StepEditorContent` component to utilize the new hook, enhancing readability and maintainability.
- Streamlined the handling of existing content and AI insertion logic, improving user experience during content generation.
- Removed redundant imports and functions, optimizing the component structure.
